Make structural_parser internal
